Move tenant authentication to new module tenant-auth, to override internally
