+++ pom.xml
@@ -120,6 +120,7 @@
         <module>standalone-container</module>
         <module>statistics</module>
         <module>storage</module>
+        <module>tenant-auth</module>
         <module>tenant-base</module>
         <module>tenant-cd</module>
         <module>testutil</module>
