+++ tenant-auth/README.md
@@ -1,0 +1,1 @@
+# Utilities that authenticate users to the hosted Vespa API, or to hosted Vespa applications.
