+++ tenant-auth/src/test/java/ai/vespa/hosted/auth/AuthenticatorTest.java
@@ -1,0 +1,5 @@
+package ai.vespa.hosted.auth;
+
+public class AuthenticatorTest {
+
+}
