+++ tenant-cd/pom.xml
@@ -40,6 +40,12 @@
 
         <dependency>
             <groupId>com.yahoo.vespa</groupId>
+            <artifactId>tenant-auth</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.yahoo.vespa</groupId>
             <artifactId>hosted-api</artifactId>
             <version>${project.version}</version>
         </dependency>
